// src/lesson_3/typescript/types.d.ts
export type TStatus = 'idle' | 'loading' | 'succeeded' | 'failed';

export type TError = string | null | Error;

export interface IReactions {
  thumbsUp: number;
  wow: number;
  heart: number;
  rocket: number;
  coffee: number;
}

// Generated by https://quicktype.io

export interface JSONPosts {
  userId: number;
  id: number;
  title: string;
  body: string;
}

// Generated by https://quicktype.io

export interface JSONUsers {
  id: number;
  name: string;
  username: string;
  email: string;
  address: Address;
  phone: string;
  website: string;
  company: Company;
}

export interface Address {
  street: string;
  suite: string;
  city: string;
  zipcode: string;
  geo: Geo;
}

export interface Geo {
  lat: string;
  lng: string;
}

export interface Company {
  name: string;
  catchPhrase: string;
  bs: string;
}